% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_value.R
\name{get_final_value}
\alias{get_final_value}
\title{Compute the final value of an investment project.}
\usage{
get_final_value(
  n_per,
  v_0,
  exp_return,
  compound,
  x_yearly = 0,
  immediate = TRUE
)
}
\arguments{
\item{n_per}{Number of periods (years) over which the final value should be
calculated.}

\item{v_0}{Initial investment capital.}

\item{exp_return}{Average expected return per year (in percent).}

\item{x_yearly}{Yearly annuity. Default = 0.}

\item{immediate}{If TRUE, immediate annuities (1st annuity on the 1st of January).
Otherwise, ordinary annuities (1st annuity on the 31st of December).}
}
\value{
- `FINAL_VALUE`: Investment final value time serie.
}
\description{
Function computing the final of an investment project.
}
\examples{
# Final value of immediate annuities with first capital and yearly expected of
# 10\% during 30 years.
get_final_value(
  n_per = 30,
  v_0 = 5000,
  exp_return = 10,
  compound = TRUE,
  x_yearly = 12000,
  immediate = TRUE
)

}
\author{
[Layal Christine Lettry](mailto:layal.lettry@gmail.com)
}
